+++ dart/constraint/SoftContactConstraint.hpp
@@ -215,10 +215,10 @@ private:
   double mRestitutionCoeff;
 
   /// Local body jacobians for mBodyNode1
-  common::aligned_vector<Eigen::Vector6d> mJacobians1;
+  std::vector<Eigen::Vector6d> mJacobians1;
 
   /// Local body jacobians for mBodyNode2
-  common::aligned_vector<Eigen::Vector6d> mJacobians2;
+  std::vector<Eigen::Vector6d> mJacobians2;
 
   /// Contact normal expressed in body frame of the first body node
   Eigen::Vector3d mBodyDirection1;